requirements.txt
django
channels
channels_redis
django-redis
daphne
pyyaml
openpyxl